auto_processor: move the 'add_command' decorator into a new 'decorators' module.
